Organize the code.
